# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /driver/account:
    get:
      operationId: getDriverAccount
      tags: [driver]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Driver profile and operational state, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
    patch:
      operationId: updateDriverAccount
      tags: [driver]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: false, properties: { fullName: { type: string }, area: { type: string }, vehicleType: { type: string } } } } }
      responses:
        "200": { description: Updated allowed profile fields, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
  /driver/earnings:
    get:
      operationId: getDriverEarnings
      tags: [driver]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Immutable earnings snapshots, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
  /driver/documents:
    get:
      operationId: listDriverDocuments
      tags: [driver]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Authorized onboarding document history, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
...
  /onboard/driver:
    post:
      operationId: onboardDriver
      tags: [auth]
      summary: Submit driver onboarding application
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverOnboard"
      responses:
        "200":
          description: Application saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDriverDocuments
      tags: [auth]
      summary: Append new driver document versions
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DriverDocumentUpdate" } } }
      responses:
        "200": { description: Documents updated, content: { application/json: { schema: { $ref: "#/components/schemas/OnboardResult" } } } }
...
  /admin/drivers:
    get:
      operationId: listDriverApplications
      tags: [auth]
      summary: List submitted driver applications
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: reuploaded, in: query, schema: { type: boolean } }
        - { name: sort, in: query, schema: { type: string, enum: [created, reuploaded] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverApplicationList"
...
  /admin/drivers/{id}:
    get:
      operationId: getDriverApplication
      tags: [admin]
      summary: Get driver application with document and decision history
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Application detail
          content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationDetail" } } }
...
  /admin/access/accounts:
    get:
      operationId: listAdminAccounts
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 50 } }
      responses:
        "200": { description: Admin accounts, content: { application/json: { schema: { $ref: "#/components/schemas/AdminPage" } } } }
    post:
      operationId: createAdminAccount
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AdminMutation" } } } }
      responses:
        "201": { description: Created account, content: { application/json: { schema: { $ref: "#/components/schemas/AdminObject" } } } }
  /admin/access/accounts/{id}:
    patch:
      operationId: updateAdminAccount
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AdminMutation" } } } }
      responses:
        "200": { description: Updated account, content: { application/json: { schema: { $ref: "#/components/schemas/AdminObject" } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AdminObject:
      type: object
      additionalProperties: true
    AdminMutation:
      type: object
      additionalProperties: true
    AdminPage:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AdminObject" } }
        page: { type: number }
        pageSize: { type: number }
        total: { type: number }
        totalPages: { type: number }
...
    DriverOnboard:
      type: object
      required: [fullName, area, vehicleType]
      properties:
        fullName:
          type: string
          minLength: 1
        area:
          type: string
          minLength: 1
        vehicleType:
          type: string
          minLength: 1
        documents:
          type: string
          description: Comma-separated ids of uploaded documents
        nationalIdFrontUrl:
          type: string
          description: Object storage path for national ID front
        nationalIdBackUrl:
          type: string
          description: Object storage path for national ID back
        criminalRecordUrl:
          type: string
          description: Object storage path for criminal record
        licenseUrl:
          type: string
          description: Object storage path for driver license
    OnboardResult:
      type: object
      required: [success, id, status]
      properties:
        success:
          type: boolean
        id:
          type: number
        status:
          type: string
...
    DriverDocumentUpdate:
      type: object
      properties:
        nationalIdFrontUrl: { type: string, minLength: 1 }
        nationalIdBackUrl: { type: string, minLength: 1 }
        criminalRecordUrl: { type: string, minLength: 1 }
        licenseUrl: { type: string, minLength: 1 }
...
    DriverApplication:
      type: object
      required: [id, userId, fullName, area, vehicleType, status]
      properties:
        id:
          type: number
        userId:
          type: number
        fullName:
          type: string
        area:
          type: string
        vehicleType:
          type: string
        documents:
          type: ["string", "null"]
        nationalIdFrontUrl:
          type: ["string", "null"]
        nationalIdBackUrl:
          type: ["string", "null"]
        criminalRecordUrl:
          type: ["string", "null"]
        licenseUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
        rejectionReason:
          type: ["string", "null"]
        latestDocumentUploadedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DriverApplicationList:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/DriverApplication" } }
        page: { type: number }
        pageSize: { type: number }
        total: { type: number }
...
    ApplicationDocument:
      type: object
      required: [id, documentType, objectPath, uploaderUserId, version, uploadedAt, reviewStatus]
      properties:
        id: { type: number }
        documentType: { type: string }
        objectPath: { type: string, description: Private object reference; access through authorized storage endpoint only }
        uploaderUserId: { type: number }
        version: { type: number }
        uploadedAt: { type: string, format: date-time }
        reviewStatus: { type: string }
        reviewedByAdminId: { type: ["number", "null"] }
        reviewedAt: { type: ["string", "null"], format: date-time }
        reviewReason: { type: ["string", "null"] }
    ApplicationDecision:
      type: object
      required: [id, actorAdminId, fromStatus, toStatus, createdAt]
      properties:
        id: { type: number }
        actorAdminId: { type: number }
        fromStatus: { type: string }
        toStatus: { type: string }
        reason: { type: ["string", "null"] }
        requestId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ApplicationDetail:
      type: object
      required: [application, documents, decisions]
      properties:
        application: { type: object, additionalProperties: true }
        documents: { type: array, items: { $ref: "#/components/schemas/ApplicationDocument" } }
        decisions: { type: array, items: { $ref: "#/components/schemas/ApplicationDecision" } }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string